DTOs and mapstruct updated

+++ src/main/java/com/ea544/blogproject/Services/UserService.java
@@ -6,13 +6,11 @@ import com.ea544.blogproject.Repo.UserRepo;
 import com.ea544.blogproject.model.entity.Comment;
 import com.ea544.blogproject.model.entity.Post;
 import com.ea544.blogproject.model.entity.UserEntity;
-import jakarta.persistence.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.Map;
 import java.util.Objects;
-import java.util.Optional;
 import java.util.function.Consumer;
 
 @Service
@@ -27,11 +25,10 @@ public class UserService extends BaseService<UserEntity, UserRepo> {
         _commentRepo = commentRepo;
     }
 
-    public void addComment(Map<String, Objects> payload) {
+    public void addComment(String username, Map<String, Objects> payload) {
         Integer tempPostId = Integer.parseInt(payload.get("postid").toString());
-        String tempUserName = payload.get("username").toString();
         String tempComment = payload.get("comment").toString();
-        UserEntity owner = _repo.findByEmailStartingWith(tempUserName);
+        UserEntity owner = _repo.findByEmailStartingWith(username);
         Post post = _postRepo.findById(tempPostId).orElseThrow();
         Comment comment = new Comment();
         comment.setPost(post);
@@ -39,20 +36,24 @@ public class UserService extends BaseService<UserEntity, UserRepo> {
         _commentRepo.save(comment);
     }
 
-    public void updateComment(int id, Map<String, Object> payload) {
+    public void updateComment(String username, int id, Map<String, Object> payload) {
         String content = payload.get("content").toString();
-        Optional<Comment> optComment = _commentRepo.findById(id);
-        if (optComment.isPresent()) {
-            Comment comment = optComment.get();
+        Comment comment = _commentRepo.findById(id).get();
+
+        if (comment.getCommentOwner().getEmail().startsWith(username)) {
             comment.setContent(content);
             _commentRepo.save(comment);
         } else {
-            throw new EntityNotFoundException("this comment doesnt exist");
+            throw new RuntimeException("this comment is not for that user");
         }
     }
 
-    public void deleteComment(int id) {
-        _commentRepo.deleteById(id);
+    public void deleteComment(String username, int id) {
+        if (_commentRepo.findById(id).get().getCommentOwner().getEmail().startsWith(username)) {
+            _commentRepo.deleteById(id);
+        } else {
+            throw new RuntimeException("the user is not owner of this comment");
+        }
     }
 
     public void vote(int postId, Consumer<Post> action) {
